backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-cloud-firestore==2.13.1
pydantic==2.5.2
python-multipart==0.0.6
